renamed path from django_active_campaign to dj_active_campaing to match package name
