+++ develop/develop/urls.py
@@ -18,5 +18,5 @@ from django.urls import include, path, re_path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('django_active_campaign.urls') ),
+    path('', include('dj_active_campaign.urls') ),
 ]
