--- django_active_campaign/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path, re_path
-from django_active_campaign import views
-
-urlpatterns = [
-    # path('', views.IndexView.as_view(), name='django_active_campaign-index'),
-    # path('simple/', views.PackageView.as_view(), name='django_active_campaign-simple'),
-    # path('simple/<slug:slug>/', views.VersionView.as_view(), name='django_active_campaign-simple-version'),
-]
-
